Subscribe a listener to the store that reads the current state
The result will be that a JSON-serialized snapshot of the state is sent over all active Socket.io connections.

*We are now publishing the whole state to everyone whenever any changes occur. This may end up causing a lot of data transfer. One could think of various ways of optimizing this (e.g. just sending the relevant subset, sending diffs instead of snapshots...), but this implementation has the benefit of being easy to write, so we'll just use it for our example app.*

In addition to sending a state snapshot whenever state changes, it will be useful for clients to immediately receive the current state when they connect to the application. That lets them sync their client-side state to the latest server state right away.

We can listen to 'connection' events on our Socket.io server. We get one each time a client connects. In the event listener we can emit the current state right away