Add ability to receive remote redux actions and start a vote from a stub in index.js
This is where we start to go beyond "regular Redux", since we are now essentially accepting remote actions into our store. However, the Redux architecture makes it remarkably easy to do: Since actions are just JavaScript objects, and JavaScript objects can easily be sent over the network, we immediately got a system with which any number of clients can participate in voting. That's no small feat!

There are some obvious security considerations here. We're allowing any connected Socket.io client to dispatch any action into the Redux store.

In most real-world cases, there should be some kind of firewall here, probably not dissimilar to the one in the Vert.x Event Bus Bridge. Apps that have an authentication mechanism should also plug it in here.

Our server now operates essentially like this:

A client sends an action to the server.
The server hands the action to the Redux Store.
The Store calls the reducer and the reducer executes the logic related to the action.
The Store updates its state based on the return value of the reducer.
The Store executes the listener function subscribed by the server.
The server emits a 'state' event.
All connected clients - including the one that initiated the original action - receive the new state.