Add middleware access to socket to actually send a given action to server
There is one major problem we have with this though: When we get the state update from the server and dispatch the SET_STATE action, it also goes right back to the server. Although the server does nothing with that action, its listener is still triggered, causing a new SET_STATE. We have created an infinite loop! That's certainly not good.

It is not appropriate for the remote action middleware to send each and every action to the server. Some actions, such as SET_STATE, should just be handled locally in the client. We can extend the middleware to only send certain actions to the server. Concretely, we should only send out actions that have a {meta: {remote: true}} property attached: